requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.5
django-embed-video==1.1.2
django-modeltranslation==0.12.1
django-rosetta==0.7.13
django-solo==1.1.2
idna==2.6
microsofttranslator==0.8
MySQL-python==1.2.5
polib==1.0.8
pytz==2017.2
requests==2.18.4
six==1.11.0
urllib3==1.22
